CV/requirements.txt
opencv-python~=4.1.2.30
torch~=1.6.0
numpy~=1.16.1
torchvision~=0.7.0
Pillow~=6.2.1
grpcio-tools